refactor: 구독하기 버튼 생성 함수로 분리 (#3)
